move all functions to lib
